Add asm code for NoneZeroCount and refine related code
