+++ codec/common/arm/deblocking_neon.S
@@ -834,17 +834,12 @@ WELS_ASM_FUNC_END
 
 
 WELS_ASM_FUNC_BEGIN WelsNonZeroCount_neon
-
-    vld1.64 {d0-d2}, [r0]
-
-    vceq.s8 q0, q0, #0
-    vceq.s8 d2, d2, #0
-    vmvn    q0, q0
-    vmvn    d2, d2
-    vabs.s8 q0, q0
-    vabs.s8 d2, d2
-
-    vst1.64 {d0-d2}, [r0]
+    mov       r1, #1
+    vdup.8    q2, r1
+    vld1.64   {d0,d1,d2}, [r0]
+    vmin.s8   q0, q0, q2
+    vmin.s8   d2, d2, d4
+    vst1.64   {d0,d1,d2}, [r0]
 WELS_ASM_FUNC_END
 
 #ifdef __APPLE__
