+++ codec/decoder/core/arm/block_add_neon.S
@@ -99,18 +99,6 @@
 //  }
 .endm
 #endif
-// r0    int8_t* non_zero_count,
-WELS_ASM_FUNC_BEGIN SetNonZeroCount_neon
-    vld1.64 {d0-d2}, [r0]
-    vceq.s8 q0, q0, #0
-    vceq.s8 d2, d2, #0
-    vmvn    q0, q0
-    vmvn    d2, d2
-    vabs.s8 q0, q0
-    vabs.s8 d2, d2
-    vst1.64 {d0-d2}, [r0]
-WELS_ASM_FUNC_END
-
 
 //  uint8_t *pred, const int32_t stride, int16_t *rs
 WELS_ASM_FUNC_BEGIN IdctResAddPred_neon
