+++ codec/decoder/core/arm64/block_add_aarch64_neon.S
@@ -100,20 +100,6 @@
 //  }
 .endm
 #endif
-// x0    int8_t* non_zero_count,
-WELS_ASM_AARCH64_FUNC_BEGIN SetNonZeroCount_AArch64_neon
-    mov x1, x0
-    ld1 {v0.16b}, [x1], #16
-    ld1 {v1.8b}, [x1]
-    cmeq v0.16b, v0.16b, #0
-    cmeq v1.8b, v1.8b, #0
-    mvn  v0.16b, v0.16b
-    mvn  v1.8b, v1.8b
-    abs  v0.16b, v0.16b
-    abs  v1.8b, v1.8b
-    st1 {v0.16b}, [x0], #16
-    st1 {v1.8b}, [x0]
-WELS_ASM_AARCH64_FUNC_END
 
 //  uint8_t *pred, const int32_t stride, int16_t *rs
 WELS_ASM_AARCH64_FUNC_BEGIN IdctResAddPred_AArch64_neon
